move enum module to enums to avoid confusion
