+++ airflow/dag/serialization/serialized_dag.py
@@ -19,7 +19,7 @@
 
 """DAG serialization with JSON."""
 
-from airflow.dag.serialization.enum import DagAttributeTypes as DAT, Encoding
+from airflow.dag.serialization.enums import DagAttributeTypes as DAT, Encoding
 from airflow.dag.serialization.json_schema import make_dag_schema
 from airflow.dag.serialization.serialization import Serialization
 from airflow.models import DAG
